+++ src/commands/index.ts
@@ -3,3 +3,5 @@ export * as help from "./help";
 export * as mark from "./mark";
 export * as page from "./page";
 export * as invoke from "./invoke";
+export * as lorem from "./lorem";
+export * as sort from "./sort";