+++ src/stores/command.ts
@@ -36,6 +36,10 @@ export const commandList = [
   { value: "rename", desc: "Rename current page.", wait: true },
   { value: "undo", desc: "Undo last edit.", wait: false },
   { value: "redo", desc: "Redo last edit.", wait: false },
+  { value: "lorem", desc: "Random generate blocks.", wait: true },
+  { value: "lorem-ipsum", desc: "Random generate blocks.", wait: true },
+  // { value: "sort", desc: "Sort blocks a-z.", wait: false },
+  // { value: "rsort", desc: "Sort blocks z-a.", wait: false },
   {
     value: "go ",
     desc: "Create new page or go to existed page.",